Oliver's updates
Reordered the parameters to new and delete. There are enough explicit rules about the
first and second parameter types for valid new and delete operators that placing the
type_identity first that I think trying to use a different order just won't fly.

Added explicit examples for new[] and delete[] as their absence might imply that they
aren't covered. We may want to just have an additional addendum saying something to the
effect of "and so on for the aligned and sized variants".

I've added a note/design choice - for this to be usable to kalloc_type we need to be
able to specify general operators new and delete, so they need to be templatized. The
rules for resolving operator delete preclude picking up the templated delete as a usual
delete operator which means that the untyped delete would be selected for at least the
virtual deleting destructor.